review changes

+++ app/Http/Controllers/OrderController.php
@@ -9,6 +9,8 @@ use Carbon\Carbon;
 Use App\CartDetail;
 use Exception;
 use App\Http\Requests\CheckOutFormValidation;
+use Illuminate\Support\Facades\DB;
+use NewException;
 
 class OrderController extends Controller
 {
@@ -17,28 +19,45 @@ class OrderController extends Controller
         try
         {
             $product=ProductDetail::productIddetail($Id);
-            return view('checkout',['product'=>$product]);
+            if(empty($product))
+            {
+                throw New Exception('Product not found');
+            }
         }
         catch(Exception $e)
         {
+            $array = [ "error"=>$e->getMessage() ];
+            return redirect('error')->withInput()->withErrors($array);
+        }
+        return view('checkout',['product'=>$product]);
+    }
+    function orderProduct(CheckOutFormValidation $request,$Id)
+    {
+        try{
+            DB::beginTransaction();
+            $request->validate();
+            $product=ProductDetail::productIddetail($Id);
+            if(empty($product))
+            {
+                throw New Exception('Product not found');
+            }
+            $price=$product[0]->price-($product[0]->price*$product[0]->discount/100);
+            OrderDetail::addOrder( [ 'name' =>$request->name,'address' =>$request->address,'pincode' =>$request->pincode,
+                'city' =>$request->city,'state' =>$request->state,'price' =>$price,'payment_mode' =>$request->cash_payment,
+                'customer_phone' =>session('active_user'),'phone_no' =>$request->phone_number,'product_id'=>$Id,'added_on'=>Carbon::now() ] );
+
+            ProductDetail::updateProduct($Id,['quantity'=>$product[0]->quantity-1]);
+            DB::Commit();
+        }
+        catch(Exception $e)
+        {
+            DB::rollback();
             $array = [
                 "error"=>$e->getMessage()
             ];
             return redirect('error')->withInput()->withErrors($array);
         }
-    }
-    function orderCheck(CheckOutFormValidation $request,$Id)
-    {
-
-        $request->validate();
-        $product=ProductDetail::productIddetail($Id);  
-        $price=$product[0]->price-($product[0]->price*$product[0]->discount/100);
-
-        OrderDetail::addOrder([ 'name' =>$request->name,'address' =>$request->address,'pincode' =>$request->pincode,
-        'city' =>$request->city,'state' =>$request->state,'price' =>$price,'payment_mode' =>$request->cash_payment,
-        'customer_phone' =>session('active_user'),'phone_no' =>$request->phone_number,'product_id'=>$Id,'added_on'=>Carbon::now()]);
-
-        ProductDetail::updateProduct($Id,['quantity'=>$product[0]->quantity-1]);
+        
         return redirect('order_successful')->with('success',"Done!!");
     }
     function orderHistory()
@@ -46,27 +65,23 @@ class OrderController extends Controller
         try
         {
             $orders=OrderDetail::ordersproductdetails(session('active_user'));
-            return view('order_history',['orders'=>$orders]);
         }
         catch(Exception $e)
         {
-            $array = [
-                "error"=>$e->getMessage()
-            ];
+            $array = [ "error"=>$e->getMessage() ];
             return redirect('error')->withInput()->withErrors($array);
         }
+        return view('order_history',['orders'=>$orders]);
     }
-    function ordersCheckout(CheckOutFormValidation $request)
+    function orderProducts(CheckOutFormValidation $request)
     {
         $request->validate();
         try{
-            $productsIdQuantity= CartDetail::productidQuantityId(session('active_user'));
+            $productsIdQuantity= CartDetail::productDetail(session('active_user'),['product_id','quantity']);
         }
         catch(Exception $e)
         {
-            $array = [
-                "error"=>$e->getMessage()
-            ];
+            $array = [ "error"=>$e->getMessage() ];
             return redirect('error')->withInput()->withErrors($array);
         }
         foreach($productsIdQuantity as $productIdQuantity )
@@ -78,55 +93,60 @@ class OrderController extends Controller
                 $quantity--;
                 $product=ProductDetail::productIddetail($productId,1);
                 $price =$product[0]->price-($product[0]->price*$product[0]->discount/100);
-                $orderTable=new OrderDetail;
                 if(($product[0]->quantity-1)>=0)
                 {
-                    OrderDetail::addOrder([ 'name' =>$request->name,'address' =>$request->address,'pincode' =>$request->pincode,
-                    'city' =>$request->city,'state' =>$request->state,'price' =>$price,'payment_mode' =>$request->cash_payment,
-                    'customer_phone' =>session('active_user'),'phone_no' =>$request->phone_number,'product_id'=>$productId,'added_on'=>Carbon::now()]);
-                    ProductDetail::updateProduct($productId,['quantity'=>$product[0]->quantity-1]);
+                    try
+                    {
+                        DB::beginTransaction();
+                        OrderDetail::addOrder([ 'name' =>$request->name,'address' =>$request->address,'pincode' =>$request->pincode,
+                            'city' =>$request->city,'state' =>$request->state,'price' =>$price,'payment_mode' =>$request->cash_payment,
+                            'customer_phone' =>session('active_user'),'phone_no' =>$request->phone_number,'product_id'=>$productId,'added_on'=>Carbon::now()]);
+                        
+                        ProductDetail::updateProduct($productId,['quantity'=>$product[0]->quantity-1]);
+                        CartDetail::deleteCartItems(session('active_user'));
+                        DB::Commit();
+                    }
+                    catch(Exception $e)
+                    {
+                        DB::rollback();
+                        $array = [ "error"=>$e->getMessage() ];
+                        return redirect('error')->withInput()->withErrors($array);
+                    }
                 }
             }
         }
-        CartDetail::deleteCartItems(session('active_user'));
         return redirect('order_successful')->with('success',"Done!!");
     }
-     function orderReceived()
-    {
-        try
-        {
-            $orders=OrderDetail::ordersproductdetails(session('active_user'));
-            return view('order_receive',['products'=>$orders]);
-        }
-        catch(Exception $e)
-        {
-            $array = [
-                "error"=>$e->getMessage()
-            ];
-            return redirect('error')->withInput()->withErrors($array);
-        }
-    }
+
     function cancelOrder($id)
      {
          try
          {
+            DB::beginTransaction();
             $orderProductid=OrderDetail::productId($id);
+            if(empty($orderProductid))
+            {
+                throw New Exception('Order not found');
+            }
             $productIdQuantity= ProductDetail::productQuantity($orderProductid);
             $quantity=$productIdQuantity[0]->quantity;
             OrderDetail::deleteOrder($id);
             ProductDetail::updateProduct($orderProductid,['quantity'=>$quantity+1]);
+            DB::Commit();
          }
          catch(Exception $e)
          {
-            $array = [
-                "error"=>$e->getMessage()
-            ];
+             DB::rollback();
+            $array = [ "error"=>$e->getMessage() ];
             return redirect('error')->withInput()->withErrors($array);
          }
         return redirect('order_history');
      }
     function status()
     {
+        //$order=new OrderDetail;
+        //foreach($order->product() as $order)
+        //dd($order->product()->where('orderId',1)) ;
         return view('order_successful');
     }
